requirements.txt
pandas
psycopg2-binary
requests
python-dotenv
sqlalchemy
gspread
gspread-dataframe
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
matplotlib
pillow
